Setting Authorization in axios interceptor instead of passing from UI component

+++ src/Features/Auth/authSlice.js
@@ -1,6 +1,5 @@
 import { networkService } from "../../main";
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
-import { jwtDecode } from "jwt-decode";
 import axios from "axios";
 
 const initialState = {
@@ -44,8 +43,8 @@ export const login = createAsyncThunk("auth/login", async (form, thunkApi) => {
 });
 
 export const update = createAsyncThunk("auth/update", async (data, thunkApi) => {
-	const { authToken: token, localData: form } = data;
-	const user = jwtDecode(token);
+	const { localData: form } = data;
+	const user = thunkApi.getState().auth.user;
 	try {
 		const fd = new FormData();
 		form.firstName && fd.append("firstName", form.firstName);
@@ -61,7 +60,6 @@ export const update = createAsyncThunk("auth/update", async (data, thunkApi) => 
 		form.deleteProfileImage && fd.append("deleteProfileImage", form.deleteProfileImage);
 
 		const res = await networkService.updateUser({
-			authToken: token,
 			userId: user._id,
 			form: fd,
 		});
@@ -78,14 +76,11 @@ export const update = createAsyncThunk("auth/update", async (data, thunkApi) => 
 	}
 });
 
-export const deleteUser = createAsyncThunk("auth/delete", async (data, thunkApi) => {
-	const user = jwtDecode(data);
+export const deleteUser = createAsyncThunk("auth/delete", async (_, thunkApi) => {
+	const user = thunkApi.getState().auth.user;
 
 	try {
-		const res = await networkService.deleteUser({
-			authToken: data,
-			userId: user._id,
-		});
+		const res = await networkService.deleteUser({ userId: user._id, });
 		return res.data;
 	} catch (error) {
 		if (error.response && error.response.data) {
